# src/openapi/spec/virtual-train-api.yaml
openapi: 3.0.1
info:
  title: Virtual Train API
  description: Virtual Train API
  version: '1.0'
servers:
  - url: https://gateway.apiportal.ns.nl/virtual-train-api
paths:
...
  /vehicle:
    get:
      tags:
        - Vehicle
      summary: getVehicles
      description: Information about all vehicles
      operationId: getVehicles
      parameters:
        - name: lat
          in: query
          description: Format - float. Latitude coordinate
          schema:
            type: number
            format: float
            default: 0
        - name: lng
          in: query
          description: Format - float. Longitude coordinate
          schema:
            type: number
            format: float
            default: 0
        - name: radius
          in: query
          description: Format - int32. Radius around the given coordinates in meters.
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Format - int32. Maximum number of vehicles to return.
          schema:
            type: integer
            format: int32
        - name: route
          in: query
          description: Format - int32. Route to filter.
          schema:
            type: integer
            format: int32
        - name: features
          in: query
          description: Type of material to return. If the materieel value is used the material number of the train will be returned.
          schema:
            enum:
              - bus
              - materieel
              - trein
            type: string
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
        '400':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
        '500':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
...
  /api/vehicle:
    get:
      tags:
        - Vehicle
      summary: getVehicles_1
      description: Information about all vehicles
      operationId: getVehicles_1
      parameters:
        - name: lat
          in: query
          description: Format - float. Latitude coordinate
          schema:
            type: number
            format: float
            default: 0
        - name: lng
          in: query
          description: Format - float. Longitude coordinate
          schema:
            type: number
            format: float
            default: 0
        - name: radius
          in: query
          description: Format - int32. Radius around the given coordinates in meters.
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Format - int32. Maximum number of vehicles to return.
          schema:
            type: integer
            format: int32
        - name: route
          in: query
          description: Format - int32. Route to filter.
          schema:
            type: integer
            format: int32
        - name: features
          in: query
          description: Type of material to return. If the materieel value is used the material number of the train will be returned.
          schema:
            enum:
              - bus
              - materieel
              - trein
            type: string
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
        '400':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
        '500':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treinen'
...
components:
  schemas:
    MaterieelDeel:
      type: object
      properties:
        materieelDeel:
          type: integer
          format: int32
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        snelheid:
          type: number
          format: float
        richting:
          type: number
          format: float
    Trein:
      type: object
      properties:
        ritId:
          type: string
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        snelheid:
          type: number
          format: float
        richting:
          type: number
          format: float
        horizontaleNauwkeurigheid:
          type: number
          format: float
        type:
          type: string
        materieelDeelList:
          type: array
          items:
            $ref: '#/components/schemas/MaterieelDeel'
        treinCloneWithMaterieel:
          $ref: '#/components/schemas/Trein'
    Treinen:
      type: object
      properties:
        treinen:
          type: array
          items:
            $ref: '#/components/schemas/Trein'
...
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      name: Ocp-Apim-Subscription-Key
      in: header
    apiKeyQuery:
      type: apiKey
      name: subscription-key
      in: query
security:
  - apiKeyHeader: [ ]
  - apiKeyQuery: [ ]